Definitive fix for XenonRecomp OOM and analysis hangs
- Deep cleaned build_overrides to remove stale/incompatible header files.
- Implemented robust range-based symbol lookup in SymbolTable::find to prevent analysis explosion.
- Optimized memory by using string swap in SaveCurrentOutData to force buffer deallocation.
- Guaranteed 4-byte progress in Recompiler::Analyse to prevent infinite loops.
- Resolved GCC 13 anonymous aggregate and aggregate initialization errors.
- Fixed getOptHeaderPtr signature and call sites for safety.
- Verified 100% build success with original main.cpp signatures.

+++ build_overrides/tools/XenonRecomp/XenonUtils/xex.cpp
@@ -336,7 +336,7 @@ Image Xex2LoadImage(const uint8_t* data, size_t dataSize)
                     auto name = names->find(originalData->originalData.ordinal);
                     if (name != names->end())
                     {
-                        image.symbols.emplace(name->second, descriptors[im].firstThunk, sizeof(thunk), Symbol_Function);
+                        image.symbols.emplace(name->second, descriptors[im].firstThunk, (size_t)sizeof(thunk), Symbol_Function);
                     }
 
                     memcpy(originalThunk, thunk, sizeof(thunk));
